+++ .test-infra/jenkins/job_PostCommit_Python_ValidatesRunner_Dataflow_V2.groovy
@@ -37,12 +37,12 @@ PostcommitJobBuilder.postCommitJob('beam_PostCommit_Py_VR_Dataflow_V2', 'Run Pyt
         gradle {
           rootBuildScriptDir(commonJobProperties.checkoutDir)
           // TODO: Enable following tests after making sure we have enough capacity.
-          // tasks(':sdks:python:test-suites:dataflow:py35:validatesRunnerBatchTests')
           // tasks(':sdks:python:test-suites:dataflow:py36:validatesRunnerBatchTests')
-          tasks(':sdks:python:test-suites:dataflow:py37:validatesRunnerBatchTests')
-          // tasks(':sdks:python:test-suites:dataflow:py35:validatesRunnerStreamingTests')
-          // tasks(':sdks:python:test-suites:dataflow:py36:validatesRunnerStreamingTests')
-          tasks(':sdks:python:test-suites:dataflow:py37:validatesRunnerStreamingTests')
+          // tasks(':sdks:python:test-suites:dataflow:py37:validatesRunnerBatchTests')
+          tasks(':sdks:python:test-suites:dataflow:py38:validatesRunnerBatchTests')
+          // tasks(':sdks:python:test-suites:dataflow:py36:validatesRunnerBatchTests')
+          // tasks(':sdks:python:test-suites:dataflow:py37:validatesRunnerStreamingTests')
+          tasks(':sdks:python:test-suites:dataflow:py38:validatesRunnerStreamingTests')
           switches('-PuseRunnerV2')
           commonJobProperties.setGradleSwitches(delegate)
         }